// packages/core/types/const.ts
export enum Expansion {
  BaseGame = "base",
  ProphecyOfKings = "pok",
}

export enum PlanetType {
  Home = "home",
  Normal = "normal",
  MecatolRex = "mecatolrex",
  Legendary = "legendary",
}

export enum Trait {
  Cultural = "cultural",
  Industrial = "industrial",
  Hazardous = "hazardous",
}

export enum Specialty {
  Cybernetics = "cybernetics",
  Biotic = "biotic",
  Propulsion = "propulsion",
  Warfare = "warfare",
}

export enum Wormhole {
  Alpha = "alpha",
  Beta = "beta",
  Delta = "delta",
  Gamma = "gamma",
  IonStormAlpha = "ionstormalpha",
  IonStormBeta = "ionstormbeta",
}

export enum SystemType {
  Blue = "blue",
  Red = "red",
  Green = "green",
}

export enum Anomaly {
  AsteroidField = "asteroidfield",
  Nebula = "nebula",
  Supernova = "supernova",
  GravityRift = "gravityrift",
  MuaatSupernova = "muaatsupernova",
}

export enum UniqueTile {
  MecatolRex = "mecatolrex",
  GhostsHomeSystem = "ghostshomesystem",
  GhostsSliceSystem = "ghostsslicesystem",
  MalliceStartingSide = "mallicestartingside",
  MalliceFlippedSide = "malliceflippedside",
  HopesEnd = "hopesend",
  Primor = "primor",
  Mirage = "mirage",
}

export enum Phase {
  Strategy = "strategy",
  Action = "action",
  Status = "status",
  Agenda = "agenda",
}

export enum StrategyCard {
  Leadership = "leadership",
  Diplomacy = "diplomacy",
  Politics = "politics",
  Trade = "trade",
  Construction = "construction",
  Warfare = "warfare",
  Technology = "technology",
  Imperial = "imperial",
}

export enum FactionAttributes {
  ExternalGhostSystem = 1,
}